Added status handler
Change-Id: I4ae8c000c4d03255d3b34cf196cd5c1b0bf0bd28

Added status handler to keep track of existing handlers and requests served.

Useful for logging the responses to urls and makes it easier to spot check if they have the expected return values.

Simple test src code was added to test the handle_request functionality of Status Handlers. Integration test to be added in future commit

Change-Id: I4ae8c000c4d03255d3b34cf196cd5c1b0bf0bd28